+++ tempdir/shell-test/main.c
@@ -1,0 +1,7 @@
+#include "source/h/control.h"
+
+int main(int argc, char **argv) {
+    //预留配置文件位置
+    loop(argc,argv);
+    exit(0);
+}
