+++ tempdir/shell-test/ps.txt
@@ -1,0 +1,4 @@
+  PID TTY          TIME CMD
+23009 pts/4    00:00:00 bash
+26273 pts/4    00:00:00 run
+27041 pts/4    00:00:00 ps
