+++ tempdir/shell-test/source/global.c
@@ -1,0 +1,3 @@
+#include "./h/global.h"
+
+int bufsize = 1024;
