+++ tempdir/shell-test/source/h/global.h
