--- tempdir/shell/global.c
@@ -1,3 +1,0 @@
-#include "global.h"
-
-int bufsize = 1024;
