--- tempdir/shell/main.c
@@ -1,7 +1,0 @@
-#include "cmds.h"
-
-int main() {
-    //预留配置文件位置
-    loop();
-    exit(0);
-}
